Implement strings decoding
I chose to implement the decoding in a way that the API would not
require for the input to be mutable.
The choices I considered were:
1. Every function receives a &mut &[u8] and change this reference to
   'advance' the point where the data is in the decoding process. This
   was rejected because it would complicate the interface, cause
   possible data losses that would make refactoring harder later, and would
   cause side efects in the functions parameters, which is not the best
   idea.
2. Use a vector. Really bad because there would be a necessity to
   remove the first elements when 'advancing' the data (O(N) complexity)
3. Return a new pointer to the advanced data.

I went with choice 3 as the 'least worst' one. It's good in complexity,
does not require copies, but requires a new return value in every
function, which is a bit ugly and can lead to messy code.